% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firing_rate_equations.R
\name{constant_input}
\alias{constant_input}
\title{Constant Input Current Function}
\usage{
constant_input(t, current, t_start = -Inf, t_end = Inf)
}
\arguments{
\item{t}{Numeric time vector.}

\item{current}{Current}

\item{t_start}{Initial time at which the input current is applied to the system.}

\item{t_end}{Final time at which the input current is applied to the system.}
}
\value{
Current function
}
\description{
Constant current function applied to the system:
}
\details{
\deqn{I(t) = I_0}
}
\examples{
constant_input(t, current = 3, t_start = 0, t_end = 30)(t = 1:10)
}
