% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuronsim-package.R
\docType{package}
\name{neuronsim}
\alias{neuronsim}
\title{\code{neuronsim} package}
\description{
Simulate the dynamics of neuronal ensembles using the
model of FREs and QIF neurons.
}
\details{
See the README on
\href{https://github.com/aldomann/neuronsim/}{GitHub}
}
