% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_dynamics}
\alias{plot_dynamics}
\title{Plot Neuronal Ensemble Dynamics}
\usage{
plot_dynamics(data, raster_data = NULL)
}
\arguments{
\item{data}{Data frame or list of data frames from FREs solution or QIF neurons simulation, result of \code{solve_fre()} and \code{simulate_qif()} respectively.}

\item{raster_data}{Raster data from QIF neurons simulation.}
}
\value{
A \code{gg} object.
}
\description{
Plot Neuronal Ensemble Dynamics
}
