% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_macro_dynamics}
\alias{plot_macro_dynamics}
\title{Plot Macroscopic Dynamics}
\usage{
plot_macro_dynamics(data)
}
\arguments{
\item{data}{FREs solution, result of \code{solve_fre()}.}
}
\value{
A \code{gg} object.
}
\description{
Plot Macroscopic Dynamics
}
