% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_macro_trajectory}
\alias{plot_macro_trajectory}
\title{Plot Firing Rate and Membrane Potential Trajectory}
\usage{
plot_macro_trajectory(data)
}
\arguments{
\item{data}{FREs solution, result of \code{solve_fre()}.}
}
\value{
A \code{gg} object.
}
\description{
Plot chaotic trajectory of the firing rate and membrane potential.
}
