% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_raster_data}
\alias{plot_raster_data}
\title{Plot Raster Data}
\usage{
plot_raster_data(data, hide_x = FALSE)
}
\arguments{
\item{data}{Raster data frame from QIF neurons simulation, result of \code{simulate_qif()}.}

\item{hide_x}{If TRUE, the \code{x} axis will not be shown.}
}
\value{
A \code{gg} object.
}
\description{
Plot Raster Data
}
