% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_qif}
\alias{simulate_qif}
\title{Simulate QIF Neurons}
\usage{
simulate_qif(params, init_state, times, input)
}
\arguments{
\item{params}{Named vector of parameters (delta, etabar, J) to pass to the FREs.}

\item{init_state}{Named vector of the initial state (r, v) for the ODE system.}

\item{times}{Time sequence for which output is wanted; the first value of times must be the initial time.}

\item{input}{Input current or stimulus function of the system.}
}
\description{
Simulate QIF Neurons
}
