% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firing_rate_equations.R
\name{sin_input}
\alias{sin_input}
\title{Sinusoidal Input Current Function}
\usage{
sin_input(t, current, frequency, t_start = -Inf, t_end = Inf)
}
\arguments{
\item{t}{Numeric time vector.}

\item{current}{Current}

\item{frequency}{The frequency of the sine wave.}

\item{t_start}{Initial time at which the input current is applied to the system.}

\item{t_end}{Final time at which the input current is applied to the system.}
}
\value{
Current function
}
\description{
Periodic current function applied to the system:
}
\details{
\deqn{I(t) = I_0 \sin(\omega t)}
}
\examples{
sin_input(t, current = 3, frequency = pi/4)(t = 1:10)
}
