% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_fre.R
\name{solve_fre}
\alias{solve_fre}
\title{Solve Firing-Rate Equations}
\usage{
solve_fre(params, current, init_r, init_v, times, method = c("rk4", "euler"))
}
\arguments{
\item{params}{Parameters to pass to the FREs.}

\item{current}{Input current of the system.}

\item{init_r}{The initial (state) value of r(t) for the ODE system.}

\item{init_v}{The initial (state) value of r(t) for the ODE system.}

\item{times}{Time sequence for which output is wanted; the first value of times must be the initial time.}

\item{method}{The integrator to use. The default integrator used is rk4.}
}
\value{
Solution of FRE equations as a data frame
}
\description{
Solve Firing-Rate Equations
}
