% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_dynamics}
\alias{plot_dynamics}
\title{Plot Neuronal Ensemble Dynamics}
\usage{
plot_dynamics(data, raster_data = NULL)
}
\arguments{
\item{data}{FREs solution, result of \code{solve_fre()}.}
}
\value{
A \code{gg} object.
}
\description{
Plot Neuronal Ensemble Dynamics
}
