% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_input_current}
\alias{plot_input_current}
\title{Plot Input Current}
\usage{
plot_input_current(data, hide_x = FALSE)
}
\arguments{
\item{data}{FREs solution, result of \code{solve_fre()}.}

\item{hide_x}{If TRUE, the \code{x} axis will not be shown.}
}
\value{
A \code{gg} object.
}
\description{
Plot Input Current
}
