% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firing_rate_equations.R
\name{solve_fre}
\alias{solve_fre}
\title{Solve Firing-Rate Equations}
\usage{
solve_fre(params, init_state, times, input, method = c("rk4", "euler"))
}
\arguments{
\item{params}{Named vector of parameters (delta, etabar, J) to pass to the FREs.}

\item{init_state}{Named vector of the initial state (r, v) for the ODE system.}

\item{times}{Time sequence for which output is wanted; the first value of times must be the initial time.}

\item{input}{Input current or stimulus function of the system.}

\item{method}{The integrator to use. The default integrator used is rk4.}
}
\value{
Solution of FRE equations as a data frame
}
\description{
Solve Firing-Rate Equations
}
