% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_fre.R
\name{solve_fre}
\alias{solve_fre}
\title{Title}
\usage{
solve_fre(
  fun,
  params,
  current,
  init_r,
  init_v,
  times,
  method = c("rk4", "euler")
)
}
\arguments{
\item{fun}{Function that computes the values of the derivatives in the ODE system (the model definition) at time t.}

\item{params}{Parameters to pass to fun.}

\item{current}{Input current of the system.}

\item{init_r}{The initial (state) value of r(t) for the ODE system.}

\item{init_v}{The initial (state) value of r(t) for the ODE system.}

\item{times}{Time sequence for which output is wanted; the first value of times must be the initial time.}

\item{method}{The integrator to use. The default integrator used is rk4.}
}
\value{
Solution of FRE equations as a data frame
}
\description{
Title
}
